% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_simdata.R
\name{gen_simdata}
\alias{gen_simdata}
\title{Generate simulated data set matching the format of the data set used for training the cord blood lead prediction model}
\usage{
gen_simdata(N = 200, covars = TRUE, seed = 9987)
}
\arguments{
\item{N}{Integer specifying the number of simulated observations (`default=200`).}

\item{covars}{Boolean specifying wether covariates should be simulated as well (`default=TRUE`)}

\item{seed}{Integer specifying seed value for reproducibility.}
}
\value{
This function returns a `data.frame` containing simulated dentine lead levels and covariates (if applicable).
}
\description{
Generate simulated data set matching the format of the data set used for training the cord blood lead prediction model
}
\author{
Hachem Saddiki
}
