% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_lead.R
\name{impute_pb_traj}
\alias{impute_pb_traj}
\title{Impute missing values of weekly child dentine lead measurement trajectories. The imputation method uses the `imputation()` method from the `longitudinalData` package.}
\usage{
impute_pb_traj(pb_data, method = "copyMean")
}
\arguments{
\item{pb_data}{`data.frame` containing longitudinal measurements of weekly child dentine lead with missing values.}

\item{method}{Character describing the imputation method to be used (`default='copyMean'`). For other options, see `?longitudinalData::imputation`.}
}
\value{
This function returns a `data.frame` matching the input `pb_data` with missing values replaced with imputed ones.
}
\description{
Impute missing values of weekly child dentine lead measurement trajectories. The imputation method uses the `imputation()` method from the `longitudinalData` package.
}
\author{
Hachem Saddiki
}
