% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lead.R
\name{pred_17wk_cbpb}
\alias{pred_17wk_cbpb}
\title{Main function to predict cord blood lead levels from weekly child dentine lead levels (17 weeks around birth).}
\usage{
pred_17wk_cbpb(pb_data, covars = NULL)
}
\arguments{
\item{pb_data}{\code{data.frame} containing weekly \bold{log2}-transformed child dentine lead measurements (17 weeks around birth in wide format).}

\item{covars}{\code{data.frame} of (optional) covariates that can be included in the model. These can only be: \bold{infant_sex} (binary), \bold{mother_age} (continuous), \bold{second_hand_smoke} (binary), \bold{SES_medium} (binary), and \bold{SES_high} (binary).}
}
\value{
This function returns a numeric \code{vector} containing predictions of cord blood lead concentrations in \bold{log2} scale.
}
\description{
The data set should be specified in wide format, with each column representing 1 of 17 weekly dentine lead levels in \bold{log2} scale. \cr
\itemize{
\item The columns for measurements corresponding to 8 weeks up to 1 week before birth should be labelled \bold{b208.log_8} to \bold{Pb208.log_1}.
\item The column for birth week should be labelled \bold{Pb208.log0}.
\item The columns corresponding to 1 week up to 8 weeks after birth should be labelled \bold{Pb208.log1} to \bold{Pb208.log8}.
}
}
\author{
Hachem Saddiki
}
